move store.ts
